introduce containerization

+++ odime-api/pkg/models/expense.go
@@ -2,7 +2,7 @@ package models
 
 type Expense struct {
 	ID        uint    `json:"id" gorm:"primaryKey"`
-	ReceiptID uint32  `json:"receipt_id"`
+	ReceiptID int64   `json:"receipt_id"`
 	Category  string  `json:"category"`
 	Amount    float32 `json:"amount"`
 	Timestamp int64   `json:"timestamp"`
